project_workspace/content_gen_platform/src/backend/requirements.txt
# Core
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2

# Async
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
redis==5.1.0

# Database
psycopg2-binary==2.9.9
pgvector==0.3.2

# ML / Embedding
sentence-transformers==3.0.1
torch==2.3.1

# Document parsing
pypdf==4.3.1
python-docx==1.1.2

# HTTP client
httpx==0.27.2

# Encryption
cryptography==42.0.8

# Storage (optional MinIO)
django-storages==1.14.4
boto3==1.34.160

# Utilities
gunicorn==22.0.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
